# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vacation Spots API",
        "description": "API for managing vacation spots and user wishlists.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/vacation-spots": {
            "get": {
                "tags": [
                    "Vacation Spots"
                ],
                "summary": "Retrieve a list of vacation spots",
                "operationId": "964eb51f7097e15bd4402e587d363243",
                "responses": {
                    "200": {
                        "description": "A list of vacation spots",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Beach Paradise"
                                            },
                                            "latitude": {
                                                "type": "number",
                                                "format": "float",
                                                "example": -21.453397
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "format": "float",
                                                "example": -173.696103
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Database error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Vacation Spots"
                ],
                "summary": "Create a new vacation spot",
                "operationId": "cab279c73300507072ab30fa70a4d5ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mountain Retreat"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -21.453397
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -173.696103
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vacation spot created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Mountain Retreat"
                                        },
                                        "latitude": {
                                            "type": "number",
                                            "format": "float",
                                            "example": -21.453397
                                        },
                                        "longitude": {
                                            "type": "number",
                                            "format": "float",
                                            "example": -173.696103
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Database error"
                    }
                }
            }
        },
        "/api/vacation-spots/{id}": {
            "get": {
                "tags": [
                    "Vacation Spots"
                ],
                "summary": "Retrieve a specific vacation spot",
                "operationId": "e0dce47c61c7b7e2efede2482a567bc1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the vacation spot",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vacation spot details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Beach Paradise"
                                        },
                                        "latitude": {
                                            "type": "number",
                                            "format": "float",
                                            "example": -21.453397
                                        },
                                        "longitude": {
                                            "type": "number",
                                            "format": "float",
                                            "example": -173.696103
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vacation spot not found"
                    },
                    "500": {
                        "description": "Database error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vacation Spots"
                ],
                "summary": "Update an existing vacation spot",
                "operationId": "8a738fcff2af61a1722dddbba3000663",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the vacation spot",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mountain Retreat"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -21.453397
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -173.696103
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vacation spot updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Mountain Retreat"
                                        },
                                        "latitude": {
                                            "type": "number",
                                            "format": "float",
                                            "example": -21.453397
                                        },
                                        "longitude": {
                                            "type": "number",
                                            "format": "float",
                                            "example": -173.696103
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Vacation spot not found"
                    },
                    "500": {
                        "description": "Database error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vacation Spots"
                ],
                "summary": "Delete a specific vacation spot",
                "operationId": "f1a0f54a6430689bec346c05750376ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the vacation spot to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vacation spot deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vacation Spot has been deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vacation spot not found"
                    },
                    "500": {
                        "description": "Database error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}